refactored the way constants are pulled in
