+++ constants/SchemaConstants.js
@@ -1,0 +1,12 @@
+module.exports = {
+    REPLACE_ALL: 'REPLACE_ALL',
+    ENQUEUE: 'ENQUEUE',
+    REPLACE_ENQUEUED: 'REPLACE_ENQUEUED',
+    CLEAR_ENQUEUED: 'CLEAR_ENQUEUED',
+    CLEAR_ALL: 'CLEAR_ALL',
+    AUDIOPLAYER: {
+        PLAY: 'AudioPlayer.Play',
+        STOP: 'AudioPlayer.Stop',
+        CLEAR_QUEUE: 'AudioPlayer.ClearQueue'
+    }
+};
